Finalisation de l'application et validation du parcours étudiant

+++ backend/src/main/java/com/monconcours/backend/service/OffreAbonnementService.java
@@ -1,19 +1,26 @@
 package com.monconcours.backend.service;
 
 import com.monconcours.backend.entity.OffreAbonnement;
+import com.monconcours.backend.repository.AbonnementRepository;
 import com.monconcours.backend.repository.OffreAbonnementRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+
 @Service
 public class OffreAbonnementService {
 
     private final OffreAbonnementRepository offreAbonnementRepository;
+    private final AbonnementRepository abonnementRepository;
 
-    public OffreAbonnementService(OffreAbonnementRepository offreAbonnementRepository) {
+    public OffreAbonnementService(
+            OffreAbonnementRepository offreAbonnementRepository,
+            AbonnementRepository abonnementRepository) {
+
         this.offreAbonnementRepository = offreAbonnementRepository;
+        this.abonnementRepository = abonnementRepository;
     }
 
     // CREATE
@@ -62,6 +69,38 @@ public class OffreAbonnementService {
 
     // DELETE
     public void supprimerOffre(Integer id) {
+
+        if (abonnementRepository.existsByOffreAbonnementId(id)) {
+            throw new RuntimeException(
+                    "Impossible de supprimer cette offre car elle est utilisée par un abonnement."
+            );
+        }
+
         offreAbonnementRepository.deleteById(id);
     }
+
+    public OffreAbonnement desactiverOffre(Integer id) {
+
+        OffreAbonnement offre = offreAbonnementRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Offre introuvable"));
+
+        offre.setActive(false);
+
+        return offreAbonnementRepository.save(offre);
+    }
+
+    public OffreAbonnement reactiverOffre(Integer id) {
+
+        OffreAbonnement offre = offreAbonnementRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Offre introuvable"));
+
+        offre.setActive(true);
+
+        return offreAbonnementRepository.save(offre);
+    }
+
+    public List<OffreAbonnement> getOffresActives() {
+        return offreAbonnementRepository.findByActiveTrue();
+    }
+
 }
